pyproject.toml
[project]
name = "twinscribe"
version = "0.1.0"
description = "Dual-Stream Code Documentation System with Tiered Model Architecture"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TwinScribe Team"}
]
keywords = ["documentation", "code-analysis", "llm", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    # LLM clients - using OpenRouter for all models
    "openai>=1.50.0",  # OpenRouter is OpenAI-compatible
    "anthropic>=0.40.0",  # Direct Anthropic client as fallback

    # Agent orchestration
    "langgraph>=0.2.0",

    # Static analysis
    "pycg>=0.0.7",
    "pyan3>=1.2.0",

    # Beads integration (issue tracking)
    "jira>=3.8.0",
    "atlassian-python-api>=3.41.0",

    # Data validation and models
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Async HTTP client
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",

    # Retry logic
    "tenacity>=8.2.0",

    # CLI and display
    "rich>=13.0.0",
    "click>=8.1.0",

    # YAML configuration
    "pyyaml>=6.0.0",

    # Environment variable handling
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",  # Parallel test execution
    "pytest-timeout>=2.2.0",  # Test timeouts
    "pytest-mock>=3.12.0",  # Enhanced mocking
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
    "respx>=0.20.0",  # Mock HTTP requests
    "factory-boy>=3.3.0",  # Test data factories
]
java = [
    "javalang>=0.13.0",
]
js = [
    "esprima>=4.0.0",
]

[project.scripts]
twinscribe = "twinscribe.cli:main"

[project.urls]
Homepage = "https://github.com/twinscribe/twinscribe"
Documentation = "https://github.com/twinscribe/twinscribe#readme"
Repository = "https://github.com/twinscribe/twinscribe"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/twinscribe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for component interactions",
    "slow: Tests that take more than 1 second",
    "llm: Tests that would require LLM API calls (mocked in CI)",
    "beads: Tests related to Beads/Jira integration",
    "static_analysis: Tests for static analysis functionality",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["twinscribe"]

[tool.coverage.run]
source = ["src/twinscribe"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true
skip_covered = false
precision = 2

[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101"]  # Allow assert in tests
